Extract the `getJavadoccableTrees` method into `Utils`
This is the same logic for finding positions to add Javadoc, and so should be
shared between both `AlmostJavadoc` and `NotJavadoc`.

Additionally, this fixes a bug in `AlmostJavadoc` where it'd suggest Javadoc
on trees which weren't Javadocable, e.g. local classes, by relying on the
`NotJavadoc` version of that method which contains this logic.

PiperOrigin-RevId: 759604753