Clarify the MissingCasesInEnumSwitch error message for switch expressions
The current implementation will only check switch expressions if it has a default branch that is commented as being present for skew. However, the current error message will tell users to ensure all cases are covered, or add a default branch.

This error is non-sensical in this context as a default branch does exist. Instead we can explicitly tell users to ensure all cases are handled in addition to having a default branch.

PiperOrigin-RevId: 763870229