Fix (behind a flag) a silly bug in RedundantSetterCall.
I clearly wanted "if this isn't a proto, return empty" here, and what I wrote was "if this is a lite proto, return empty". The bug would have been more obvious if GeneratedMessage extended GeneratedMessageLite (this is true if you drop the "Generated"!), as the code would have been hopelessly broken, but this isn't the case.

To add further to the pain, this is a no-op given lite proto oneof enums don't implement the marker interface. I'm seeing if I can fix that in unknown commit, but I'll need _a_ flag in place so that all the ensuing detected errors don't break.

Also this is borderline impossible to test given you can't compile lite protos into the same target as full-fat protos.

PiperOrigin-RevId: 799522744