BadInstanceof: test for interaction with pattern variable
The `BadInstanceof` check argues that the construct

    if (t() instanceof T some) {
      return some;
    }

is equivalent to a null check. That's true in the classic case where
some local stands in the place of `t()`. However, with JDK 16's
`instanceof` pattern variables we have the ability to only conditionally
declare a local for `t()`, and in that case using the prescribed null
check is a greater syntactic and semantic change. This commit adds a
test for this construct for posterity.

It may be that the example should be discouraged for the same basic
clarity reasons that underlie the `BadInstanceof` check, but then a new,
more focused check seems warranted.